Fixes #198 Add support for lists of Scalar field arguments that are already coerced
graphql-java already converts scalar types and lists of scalar types
when used as field arguments.

However the MethodFieldResolver basically serialized that object and
deserialized the value to use it as argument value while it's not really
necessary.

The serialization/deserialization also failed when the (Java) type of the
scalar is not easily serializable/deserializable (for example with
private fields)