Treat non-nullable scalars as scalars
When using a non-nullable scalar as input type
a regression was introduced by the upgrade of jackson.

A jackson exception occurs internally in the MethodFieldResolver
because the correct deserialiser/serialiser cannot be found.

Somehow this was hidden when using earlier versions of jackson
where this just worked anyhow, but due to changes in jackson this
bug is uncovered. The correct way is of course to fix this in
isScalarType to cover non-nullable scalars, and not rely on
functionality in jackson which no longer seems to be supported.